Initial `create-react-app react-enzyme-tdd` build
This is the initial project setup provided by create-react-app. Note
This does not come with Enzyme or react-router! We’ll have to add those
Later if we want them.

The only other change I made was to add the .idea folder to .gitignore
As it’s pretty noisy and there isn’t much jet brains config you 
Typically have to do in a javascript project.