Let’s add enzyme! (And add types to make linter errors go away)
Enzyme in this instance will be our component testing framework. Enzyme
Is very powerful in that it allows you to do jQuery like inspections on
Your component without having to do almost any setup whatsoever! Enzyme
Can also be used to assert on the class instance of the component as
Well which is really nice from a pure unit test perspective.