We’re still green here, but let’s refactor using enzyme
This could be considered the “refactor” stage (of Red, Green, Refactor). 
It’s not necessary that we do this, but it’s the most simple 
Introduction to enzyme I could think of and it is a pure refactor from
A test coverage perspective. 

If you’re interested in which matchers are available for the “shallow”
API, the [documentation](https://github.com/airbnb/enzyme/blob/master/docs/api/shallow.md) is pretty robust.