Let’s write our first failing test!
The purpose of this initial test is to drive out the need to pull
In react-router-dom. I’m familiar with this API so I had a good idea
Of what to write initially, but their (documentation)[https://reacttraining.com/react-router/web/guides/quick-start] is pretty good also. 

Basically what this test is saying is that on the default route “/“, 
We’d expect the router to render the landing page.

Let’s run `yarn test` to figure out what we need to do next.