Hmm.. Still red! Let’s actually write some code.
`Method “props” is only meant to be run on a single node. 0 found instead`

So it might not be exactly obvious what’s going on here, but enzyme is
Telling us the component we told it to find doesn’t actually exist. 

So now that we have a failing test describing what code we need to
Write, let’s actually change the source file to see what happens next.