Landing isn’t defined.
This is true, let’s create it!

Notice I’m not STRICTLY test driving this code since I’m writing more
That *exactly* what the failing test case says, but I personally think
It’s okay to not test drive boiler plate. The reasoning for this is 
that any test more complex than the one we have here will drive out
This boilerplate and it’s easier to write most of this when we’re 
Initially setting up this component.

This is SOMEWHAT of a controversial viewpoint, but it’s a concept I’ve
Become comfortable with over time and haven’t had many missed tests 
Over stuff like this. It’s the same argument as not testing DTOs in java. 
Obviously using some sort of code generation library like 
(create-component-app)[https://github.com/CVarisco/create-component-app]
would be a good solve to this problem, but there isn’t much boilerplate
 in react to be honest.