Let’s make it pass!
This test was pretty simple, but it also gives us a chance to look at
How our tests are set up. Notice how we’re already repeating ourselves 
on lines 8 and 14.. let’s fix that in our “refactor” stage!

Note: the point of tests is NOT to be as DRY (don’t repeat yourself) as 
possible. Repeating yourself in test isn’t the worst thing in the world
Because it lets the future reader of the test understand what you felt
Was important in the test setup or execute. 

If you try and get too fancy and extract everything into helper methods
you’ll obfuscate your intentions and make your test suite more brittle
That need be. 

However, if it’s pure test boilerplate, abstracting that boilerplate 
might actually help the reader if they have to scan hundreds of tests.
Like everything else in software engineering, there’s a balance to be
struck.