Let’s add login functionality!
Notice we moved the beforeEach to a higher scope so multiple describe
Blocks could inherit the component for free. 

Also notice the bracket syntax on lines 25 and 26. These are selectors
That describe data attributes, which we’ll be leveraging very heavily
To describe unique components on the page. 

Why data attributes vs classes or IDs? The answer is that both IDs and
Classes could get refactored based on your stylesheet, but data 
Attributes are somewhat reserved for data purposes (not that you can’t
Write a css selector that looks for a data attribute, it’s just not 
Great practice). 

Finally `.toEqual(true)` and `.toBeTruthy()` are equivalent in this
Use case. One again the goal of tests is to be readable and I personally
find this a little easier to read.