Now add the missing “user info” banner
We now have a very descriptive test failure! Now, could we just update 
The landing page to match the string, then write a second test that has
Different text to force us to make it dynamic, but honestly that’s 
Being a little too naive for my tastes. 

Let’s go ahead and assume here that the data we’re missing has to come
From a login endpoint!