Let’s mock our network call BEFORE the test runs!
Basically what we’re doing here is overriding the axios post method to
Only return a promise that resolves fake data (in this case, banana). 

This will ensure that our test will Never actually reach out to the 
Real service, and since we’re just a jest mock, we can assert on what
It gets called with! 

This gives us a very specific idea of how this integration is supposed 
To work. While this implementation does describe the code pretty 
Specifically, I tend to have a soft spot for describing external network
Calls since they are the main source of bugs in my experience.