Finally we’re able to write the bulk of the code!
Since we’ve fully described the functionality we want + put mocks in 
Place for all of the external service calls, we can now proceed with
implementation. The good news is that we can now refactor to our hearts
Content now that this code is green. 

Notice the replacement of the dummy promise data with real promise 
Data. Also notice we moved from simulating a click on the submit
Button to a simulation of the submit action on the form. This is a 
Rusty corner in enzyme that we had to work around, but I left it in 
Because it was a part of the journey. 

Finally, we had to call the update() method in the component since 
The component wasn’t updating the DOM even though we were waiting on 
The promise to resolve and the state was actually up to date.